pyproject.toml
[project]
name = "life-coach-agent"
version = "0.1.0"
description = "A Streamlit life coach agent with conversation memory and web search."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.6",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]
